requirements.txt
requests==2.26.0
cryptography==3.4.7
python-dotenv==0.19.0
